+++ resources/php_pool.rb
@@ -138,6 +138,16 @@ property :path,
     "#{node['php']['fpm_pooldir']}/#{file_name}"
   }
 
+# Ensure that the resource is applied regardless of whether we are in why_run
+# or standard mode.
+#
+# Refer to chef/chef#4537 for this uncommon syntax
+action_class do
+  def whyrun_supported?
+    true
+  end
+end
+
 action :create do
   template new_resource.path do
     path      new_resource.path