+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 php_wrapper
 ======
 # 2.0.1
-* Add whyrun support to php_composer_binary
+* Add whyrun support to php_composer_binary and php_composer_command
+* Add whyrun support to php_config and php_pool
 
 # 2.0.0
 * Initial rewrite